Change return type of unFavoriteMovie function (from Long to Int)

+++ Data/src/main/java/ir/omidtaheri/data/datasource/local/MovieLocalDataSourceInterface.kt
@@ -5,6 +5,6 @@ import kotlinx.coroutines.flow.Flow
 
 interface MovieLocalDataSourceInterface {
     suspend fun favoriteMovie(movie: FavoritedMovieDataEntity): Long
-    suspend fun unFavoriteMovie(movie: FavoritedMovieDataEntity): Long
+    suspend fun unFavoriteMovie(movie: FavoritedMovieDataEntity): Int
     fun getFavoritedMoviesList(): Flow<List<FavoritedMovieDataEntity>>
 }
